# ENH: improve discrepancy performance