… benchmarks/benchmarks/stats.py
… scipy/stats/_qmc_cy.pyx
… scipy/stats/setup.py
… scipy/stats/tests/test_qmc.py