<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
https://docs.scipy.org/doc/numpy/dev/development_workflow.html#writing-the-commit-message

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->
- [x] Did you check that the code can be distributed under a BSD license?
- [x] Are there unit tests with good code coverage?
- [x] Do all unit tests pass locally?
- [x] Do all public function have docstrings including examples?
- [x]  Does the documentation render correctly? Not sure because I have some building issue. EDIT: yes It renders correctly
- [x] Is the code style correct? I don't know if there is something for cython code's format.
- [x] Are there benchmarks? There wasn't but I added some for the discrepancy. I didn't add any for
  `update_discrepancy` though.
- [x] Is the commit message? I think so
- [x] Is the docstring of the new functionality tagged with
   ``.. versionadded:: X.Y.Z``. Not sure what it means
- [x] If compiled code is added, is it integrated correctly via ``setup.py``? I think so since I am able to run
the tests are passing

#### Reference issue
<!--Example: Closes gh-WXYZ.-->
See gh-13474. 
The new code is written in Cython in a new file `_discrepancy.pyx`. Since I don't know so much how it works with scipy, should I send a mail to the dev mailing-list to ask where to put my code ?

#### What does this implement/fix?
<!--Please explain your changes.-->
This Enhancement improves the performance of `scipy.stats.qmc.discrepancy` and `scipy.stats.qmc.update_discrepancy`. 
See additional information for performance. 
In short, the cython code achieved in average, 4 times performance's improvement compared to the scipy code. Also, this code has similar performance as the library [openturns](http://openturns.github.io/openturns/master/contents.html)

#### Additional information
To tests the performance, I run the following script : 
```python
import timeit

import numpy as np
import openturns as ot

from discrepancy_scipy import discrepancy_scipy
from discrepancy import discrepancy

np.random.seed(0)
TEST_VALUES = [
    (100, 2),
    (100, 10),
    (1000, 10),
    (1000, 100),
]
COL = [
    "old scipy time", "cython time", "openturns time", "scipy / cython",
    "openturns / cython"
]
print("Beginig")
print("||" + "|".join(f"`{colname}`" for colname in COL) + "|")
print("|" + "---|" * (len(COL) + 1))
for n, d in TEST_VALUES:
    sample = np.random.random_sample((n, d))
    ot_sample = ot.Sample(sample)

    scipy_time = np.array(timeit.repeat(
        "discrepancy_scipy(sample)",
        number=50,
        repeat=10,
        setup="from __main__ import discrepancy_scipy, " "sample",
    ))
    cython_time = np.array(timeit.repeat(
        "discrepancy(sample)",
        number=50,
        repeat=10,
        setup="from __main__ import discrepancy, " "sample",
    ))
    ot_time = np.array(timeit.repeat(
        "ot.SpaceFillingC2().evaluate(ot_sample)",
        number=50,
        repeat=10,
        setup="from __main__ import ot_sample,ot",
    ))

    print(f"|time for `n={n}` & `d={d}`", end="|")
    print(f"{scipy_time.mean():.4f}s", end="|")
    print(f"{cython_time.mean():.4f}s", end="|")
    print(f"{ot_time.mean():.4f}s", end="|")
    print(f"{scipy_time.mean() / cython_time.mean():.4f}", end="|")
    print(f"{ot_time.mean() / cython_time.mean():.4f}", end="|")
    print()
```

The results are as follow : 
||`old scipy time`|`cython time`|`openturns time`|`scipy / cython`|`openturns / cython`|
|---|---|---|---|---|---|
|time for `n=100` & `d=2`|0.0117|0.0027|0.0024|4.2846|0.8891|
|time for `n=100` & `d=10`|0.0446|0.0094|0.0099|4.7471|1.0515|
|time for `n=1000` & `d=10`|3.5753|0.9172|1.0067|3.8981|1.0976|
|time for `n=1000` & `d=100`|36.3270|9.0041|9.7423|4.0345|1.0820|

Also, I have written some benchmarks : 
![Capture d’écran 2021-02-17 à 15 04 52](https://user-images.githubusercontent.com/37664438/108218170-ad05ab00-7134-11eb-92c4-1e8ac53b143a.png)
Unfortunately, due to some building issues, I haven't been able to compare with the previous version for this particular bencharmk. I will write an issue following this PR to get some more informations.