@tupui 2021-02-17
Thank you for this PR!

Some quick notes: do not include the c files as these are cython generated and machine dependent. To answer your question, the path of the cython file is correct.