@V0lantis 2021-02-17
> Some quick notes: do not include the c files as these are cython generated and machine dependent. To answer your question, the path of the cython file is correct.

Thank you? Indeed I was wondering what to do with the c file. I removed it and force pushed my change since there is now review yet on my code

> I expected a small speed up for C2 as this was already vectorized in a proper way. Good that you compared to OT as it's C++ on their side 👍. I hope for better speed up for the rest as L2-star is particularly slow right now.

I am going to test L2-Star right away to give more information then