@tupui 2021-02-18
A quick note about comparing results with OpenTURNS. You can do it, just be aware that there is a power of 2 difference between our results. For centered discrepancy, we return c2 and they return c.