@V0lantis 2021-02-18
> For the context, I work with @V0lantis. (Some kind of light supervision).
> 
> For testing, I implement a similar computation than the discrepancy (it is not the discrepancy), and I tried to parallelize.
> 
> * [`test1.pyx`](https://perso.crans.org/leger/qad_pthreads/test1.pyx) : the reference
> * [`test2.pyx`](https://perso.crans.org/leger/qad_pthreads/test2.pyx) : the paralellized version with OpenMP. This is not allowed in scipy.
> * [`test3.pyx`](https://perso.crans.org/leger/qad_pthreads/test3.pyx) : the paralellized version with pthreads. This is quick code (and dirty). The number of threads is in argument.
> 
> Results:
> 
> ```
> Python 3.9.1 (default, Dec  8 2020, 07:51:42)
> Type 'copyright', 'credits' or 'license' for more information
> IPython 7.19.0 -- An enhanced Interactive Python. Type '?' for help.
> 
> In [1]: import numpy as np
>    ...: import test1
>    ...: import test2
>    ...: import test3
>    ...: M = np.random.uniform(size=(10000,100))
> 
> In [2]: test1.f(M)
> Out[2]: 65268094.34825731
> 
> In [3]: test2.f(M)
> Out[3]: 65268094.350631215
> 
> In [4]: test3.f(M,0)
> Out[4]: 65268094.34825731
> 
> In [5]: test3.f(M,4)
> Out[5]: 65268094.35037216
> 
> In [6]: test3.f(M,8)
> Out[6]: 65268094.350631215
> 
> In [7]: %timeit test1.f(M)
> 8.86 s ± 92.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 
> In [8]: %timeit test2.f(M)
> 2.09 s ± 347 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 
> In [9]: %timeit test3.f(M,0)
> 7.94 s ± 110 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 
> In [10]: %timeit test3.f(M,4)
> 2.59 s ± 231 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 
> In [11]: %timeit test3.f(M,8)
> 1.79 s ± 197 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> ```
> 
> Then (questions are for scipy contributors and maintainers):
> 
> * Do the usage of pthread (or similar) allowed in cython code for scipy?
> * If allowed, what is the best choice? The `pthread` C version (as I did in my test) or the `thread` lib in the std lib in C++11?
> * If allowed, do we have to produce alternative version if `pthread` not available in compilation? (seems a check like that in `scipy/fft/_pocketfft/setup.py`)
> * If allowed, do the number of thread should be provided as a argument for the user? I think yes, because when I use multiprocessing I want to disable threading.
> 
> _P.S.: It was the first time I mix cython and pthread, it was fun._

@rgommers @mdhaber Do you have some opinions on this?