@jb-leger 2021-02-18
To complete my previous post, I tried the 4th solution, use of `thread` in `C++`. This is very clean comparing to `pthread`, no pointers, memory views can be passed, it is better than `pthread` (if C++ with cython allowed). The file for testing is [`test4.pyx`].

The results sumarized:

| Implementation | nthreads | time (s) |
|---|---|---|
| [`test1.pyx`] | | `8.28` |
| [`test2.pyx`] (openmp) | 8 | `2.18` |
| [`test3.pyx`] (pthread) | 0 (no threads) | `7.60` |
| [`test3.pyx`] (pthread) | 4 | `2.21` |
| [`test3.pyx`] (pthread) | 8 | `1.42` |
| [`test4.pyx`] (C++ thread) | 0 (no threads) | `8.28` |
| [`test4.pyx`] (C++ thread) | 4 | `2.54` |
| [`test4.pyx`] (C++ thread) | 8 | `1.58` |

[`test1.pyx`]: https://perso.crans.org/qad_pthreads/test1.pyx
[`test2.pyx`]: https://perso.crans.org/qad_pthreads/test2.pyx
[`test3.pyx`]: https://perso.crans.org/qad_pthreads/test3.pyx
[`test4.pyx`]: https://perso.crans.org/qad_pthreads/test4.pyx

Therefore, C++ thread and pthread are equivalent, but code with C++ thread is
more maintainable than one with pthread.