@tupui 2021-02-18
FYI, since #8306 there is an experiment in SciPy with Pythran.