@jb-leger 2021-02-23
We (@serge-sans-paille and I) was not benchmarking the same method. I rewrite the benchmark, for considering all the methots. Note: now times are given for on function call. Here is [the script](https://perso.crans.org/leger/ana.py). And the results are following.

There is a significant improvement for `method="CD"` and for `method="MD"`, no improvement for `method="WD"` and for `method="L2-star".

## method='CD'

||`old scipy time`|`cython time`|`pythran`|`pythran1727`|`scipy/cython`|`scipy/pythran`|`scipy/pythran1727`|
|---|---|---|---|---|---|---|---|
|time for `n=100` & `d=2`|128.5µs|52.5µs|84.7µs|61.5µs|2.4465|1.5183|2.0904|
|time for `n=100` & `d=10`|530.3µs|196.3µs|350.1µs|257.6µs|2.7022|1.5148|2.0583|
|time for `n=1000` & `d=10`|73.0ms|13.0ms|24.8ms|19.4ms|5.6355|2.9410|3.7692|
|time for `n=1000` & `d=20`|141.2ms|29.8ms|48.0ms|38.5ms|4.7362|2.9393|3.6691|

## method='WD'

||`old scipy time`|`cython time`|`pythran`|`pythran1727`|`scipy/cython`|`scipy/pythran`|`scipy/pythran1727`|
|---|---|---|---|---|---|---|---|
|time for `n=100` & `d=2`|65.4µs|32.0µs|39.0µs|37.1µs|2.0403|1.6744|1.7639|
|time for `n=100` & `d=10`|292.5µs|92.7µs|147.6µs|135.4µs|3.1570|1.9818|2.1603|
|time for `n=1000` & `d=10`|69.4ms| 8.2ms|23.1ms|23.2ms|8.4492|3.0057|2.9838|
|time for `n=1000` & `d=20`|151.0ms|20.5ms|41.6ms|41.4ms|7.3774|3.6280|3.6461|

## method='MD'

||`old scipy time`|`cython time`|`pythran`|`pythran1727`|`scipy/cython`|`scipy/pythran`|`scipy/pythran1727`|
|---|---|---|---|---|---|---|---|
|time for `n=100` & `d=2`|150.0µs|46.5µs|85.5µs|60.5µs|3.2255|1.7546|2.4794|
|time for `n=100` & `d=10`|657.3µs|176.0µs|376.8µs|250.8µs|3.7354|1.7445|2.6203|
|time for `n=1000` & `d=10`|103.3ms|17.1ms|29.7ms|25.9ms|6.0380|3.4790|3.9823|
|time for `n=1000` & `d=20`|211.5ms|34.2ms|58.7ms|48.8ms|6.1881|3.6031|4.3303|

## method='L2-star'

||`old scipy time`|`cython time`|`pythran`|`pythran1727`|`scipy/cython`|`scipy/pythran`|`scipy/pythran1727`|
|---|---|---|---|---|---|---|---|
|time for `n=100` & `d=2`|284.9µs|29.6µs|38.4µs|39.6µs|9.6365|7.4095|7.1937|
|time for `n=100` & `d=10`|753.0µs|70.5µs|88.2µs|87.7µs|10.6809|8.5407|8.5839|
|time for `n=1000` & `d=10`|90.9ms| 5.8ms| 9.1ms| 9.0ms|15.6317|10.0278|10.0509|
|time for `n=1000` & `d=20`|175.5ms|13.0ms|17.0ms|16.1ms|13.4838|10.3211|10.8885|